+++ NEWS.md
@@ -1,6 +1,7 @@
 # wrappedtools 0.8.1
 - pairwise_fisher_test is now using MonteCarlo with 10^5 replications
 - faketrial dataset has new column Responder, useful in logistic regression
+- compare2qualvars has a new option to test subgroups by recoding other levels into other, default is not to do this
 
 # wrappedtools 0.8.0
 - new function surprisal (Shannon information from p-value)